+++ output/example.experiment.sol
@@ -7,7 +7,7 @@ contract Contract {
         uint64 b = 1e18;
 
 assert (a <= type(uint64).max / b); //slytherin 1
-        uint256 c = a * b;
+        uint64 c = a * b;
 
         return c;
     }
